requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.4
azure-common==1.1.27
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
Brotli==1.0.9
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
cryptography==3.4.7
cssselect2==0.4.1
dj-database-url==0.5.0
Django==3.2
django-ckeditor==6.0.0
django-filter==2.4.0
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-model-utils==4.1.1
django-polymorphic==3.0.0
django-rest-polymorphic==0.1.9
django-storages==1.11.1
djangorestframework==3.12.4
docopt==0.6.2
fonttools==4.29.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==2.10
num2words==0.5.10
paywix==1.4.0
Pillow==9.0.0
psycopg2-binary==2.9.3
pycparser==2.20
pydyf==0.1.2
PyPDF2==1.26.0
pyphen==0.12.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
reportlab==3.6.6
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tinycss2==1.1.1
urllib3==1.26.4
uuid==1.30
weasyprint==54.0
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
zopfli==0.1.9
